# .momorph/contexts/api-docs.yaml
openapi: 3.0.3
info:
  title: "SAA 2025 — Kudos API"
  description: |
    API specification for the **Kudos** subsystem of the Sun* Awards Annual 2025
    platform. Two screens back onto this surface:

    - **Viết Kudo** (screen `ihQ26W78P2`) — modal: recipient search, Danh hiệu
      & hashtag master lists, image upload, `@mention` suggestions, and Kudo
      creation.
    - **Sun* Kudos — Live board** (screen `MaZUn5xHXZ`) — public feed: HIGHLIGHT
      KUDOS carousel (top-5 hearted), ALL KUDOS cursor-paginated feed, filters
      (hashtag + department), like / un-like, personal stats, Spotlight board.

    ## Authentication
    All endpoints (except `GET /health`) require a JWT Bearer token issued by
    `POST /auth/login` (defined in the auth/login feature's spec).
    ```
    Authorization: Bearer <token>
    ```

    ## Error Responses
    All endpoints return errors in the following format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": {}
      }
    }
    ```

    ## Anonymity
    The `isAnonymous` flag on `POST /kudos` is enforced **server-side**.
    The public `GET /kudos` payload omits the author's identity when
    `isAnonymous=true`; the database still retains `author_id` for moderation.
  version: "1.0.0"
  contact:
    name: "SAA Platform Team"
    email: "saa-platform@sun-asterisk.com"
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /employees/search:
    get:
      tags: [Employees]
      summary: Search employees
      description: |
        Single endpoint for two UI surfaces:
        - **Recipient picker** — call with `ignore_caller=true` (default) to hide
          the current user from results.
        - **@mention popover** — call with `ignore_caller=false` so the caller
          may also be suggested.

        Matches `q` against `full_name` (case-insensitive, diacritic-tolerant
        prefix via pg_trgm) and `email`. Soft-deleted accounts are always
        excluded.
      operationId: searchEmployees
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Name / email query (whitespace trimmed)
        - name: ignore_caller
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: |
            When `true` (default), excludes the authenticated caller from the
            result set — use this for the Recipient field.
            When `false`, includes the caller — use this for the `@mention`
            suggestion popover.
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Matching employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ============================================
  # Departments (Phòng ban master list — Live board filter)
  # ============================================
  /departments:
    get:
      tags: [Departments]
      summary: List active departments
      description: |
        Returns every active department, sorted by `sort_order` then `code`.
        Populates the Live board Phòng ban filter dropdown (Figma `WXK5AYB_rG`).
        Soft-deleted rows are excluded.
      operationId: listDepartments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================
  # Me (Live board sidebar — personal stats)
  # ============================================
...
  /spotlight:
    get:
      tags: [Spotlight]
      summary: Spotlight data (event-wide total + top-20 recipients of the last 24h)
      description: |
        Drives the Live board Spotlight canvas (Figma `2940:14174`).

        **Data shape**:
        - `data.total` — event-wide `COUNT(*)` of published Kudos, shown as
          the big "{N} KUDOS" label (B.7.1). **Not** 24h-scoped.
        - `data.nodes[]` — up to **20** recipients ranked by `kudos_count DESC`
          within the rolling last 24 hours. Empty during quiet periods.

        **Caching**: the full response is computed server-side and cached for
        **5 minutes** (`Cache-Control: public, max-age=300, s-maxage=300`).
        `layoutVersion = {eventDayIso}:{5-min-bucket}` — clients detect changes
        to decide whether to animate a redraw.

        **Realtime complement**: clients subscribe to the Supabase Realtime
        channel on the `kudos` table for INSERT/DELETE events; those events
        only update the `total` label (debounced 500 ms) and the animated
        "recent-receivers" log inside the canvas. **Canvas nodes are NOT
        modified between 5-min polls** — the top-20 aggregate is recomputed
        server-side only.

        **Node click**: navigates to `/profile/{id}` (Kudo detail page deferred
        per plan Q-A1; node does not link to a single Kudo).
      operationId: getSpotlight
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Total + nodes + layout version
          headers:
            Cache-Control:
              schema: { type: string }
              description: "public, max-age=300, s-maxage=300"
            ETag:
              schema: { type: string }
              description: Hash of `layoutVersion`; enables 304 on unchanged layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotlightResponse"
        "304":
          description: Layout unchanged since the client's ETag
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /auth/login

  # ============================================
  # Reusable Parameters
  # ============================================
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page

  # ============================================
  # Reusable Responses
  # ============================================
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired token"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Validation failed"
              details:
                recipientId: ["Required"]
                hashtagIds: ["Must contain 1–5 items"]
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
...
    EmployeeSummary:
      type: object
      description: |
        Employee record surfaced by `/employees/search`. Fields mirror the
        `employees` master-data table (see database-schema.sql).
        On anonymous Kudo payloads the author's EmployeeSummary is NOT
        returned — the Kudo response carries pre-resolved `senderName` and
        `senderAvatarUrl` only.
      properties:
        id:
          type: integer
          format: int64
          description: Primary key of the employees master-data table
          example: 1024
        fullName: { type: string, example: "Nguyễn Văn A" }
        email: { type: string, format: email }
        employeeCode: { type: string, nullable: true, example: "SA-01234" }
        department: { type: string, nullable: true, example: "Engineering" }
        jobTitle: { type: string, nullable: true, example: "Senior Engineer" }
        avatarUrl: { type: string, format: uri, nullable: true }
    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/EmployeeSummary" }

    # --- Title (Danh hiệu) ---
...
    Department:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: "CEVC2" }
        name:
          type: string
          nullable: true
          example: "Customer Experience VC 2"
        parentId:
          type: integer
          format: int64
          nullable: true
          description: Optional hierarchy (e.g. `CEVC1 - DSV - UI/UX 1`).
        sortOrder: { type: integer }
    DepartmentListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Department" }

    # --- Me / stats ---
...
    SpotlightNode:
      type: object
      description: |
        One **top-20 recipient** positioned on the Spotlight canvas. "Top" =
        recipients ranked by `kudos_count` received **in the rolling last 24
        hours**. Identity = `employee_id` is stable across layout redraws so
        the client can animate moves; position (`x`, `y`) may shift as the
        24h window rolls over.
      required: [id, name, kudosCount, x, y]
      properties:
        id:
          type: integer
          format: int64
          description: employees.id of the recipient this node represents.
        name: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
        kudosCount:
          type: integer
          minimum: 1
          description: Kudos received by this recipient within the last 24h.
        lastReceivedAt:
          type: string
          format: date-time
          description: Timestamp of their most-recent Kudo within the 24h window (hover tooltip).
        x:
          type: number
          description: Horizontal position within the canvas (0..1 normalised).
        y:
          type: number
          description: Vertical position within the canvas (0..1 normalised).
    SpotlightResponse:
      type: object
      properties:
        data:
          type: object
          required: [total, nodes, layoutVersion, cachedAt]
          properties:
            total:
              type: integer
              minimum: 0
              description: Event-wide `COUNT(*)` of published Kudos at `cachedAt` (not 24h-scoped — this is the big B.7.1 number).
            nodes:
              type: array
              maxItems: 20
              description: Up to 20 top-recipient nodes (top `kudos_count DESC` within last 24h). May be empty during quiet periods.
              items: { $ref: "#/components/schemas/SpotlightNode" }
            layoutVersion:
              type: string
              description: "Opaque version tag `{eventDayIso}:{bucket5min}`; changes every 5 min."
              example: "2026-04-21:4712345"
            cachedAt:
              type: string
              format: date-time
              description: When the server computed this layout (TTL 5 min).